sprint5/daily-excecoes/requirements.txt
alembic==1.7.6
black==22.1.0
click==8.0.3
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
SQLAlchemy==1.4.31
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.0.3
alembic==1.7.6
black==22.1.0
click==8.0.3
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
SQLAlchemy==1.4.31
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.0.3
